Removed core library
