+++ backend/users/models.py
@@ -3,9 +3,7 @@ from django.contrib.auth.base_user import BaseUserManager
 from django.contrib.auth.models import AbstractUser
 from django.db import models
 
-from core.validators import EmailValidator
-
-from .validators import NameUserValidator
+from .validators import EmailValidator, NameUserValidator
 
 
 class UserManager(BaseUserManager):
