+++ backend/users/validators.py
@@ -10,6 +10,9 @@ from django.utils.regex_helper import _lazy_re_compile
 
 from backend.settings import (
     MAX_LEN_NAME_USER,
+    MAX_LENGTH_EMAIL_USER_PART,
+    MESSAGE_EMAIL_NOT_VALID,
+    MESSAGE_EMAIL_USER_PART_VALID,
     MESSAGE_NAME_USER_CYRILLIC,
     MESSAGE_NAME_USER_VALID,
     MIN_LEN_NAME_USER,
@@ -91,3 +94,22 @@ class PasswordRegexValidator:
             "В пароле допускаются цифры, буквы и спецсимовлы "
             "-!#$%&'*+/=?^_;():@,.<>`{}"
         ),
+
+
+class EmailValidator:
+    """
+    Проверка длинны пользовательской части email адреса.
+    """
+
+    email_max_length = MaxLengthValidator(
+        MAX_LENGTH_EMAIL_USER_PART,
+        message=MESSAGE_EMAIL_USER_PART_VALID
+    )
+
+    @classmethod
+    def validate_email(cls, value):
+        try:
+            user_part, _ = value.rsplit("@", 1)
+        except ValueError:
+            raise ValidationError(MESSAGE_EMAIL_NOT_VALID)
+        cls.email_max_length(user_part)
